Moved defences about
